Disable uploading during archiving

+++ cmd/web/handlers.go
@@ -25,6 +25,12 @@ func (app *application) createSnippetForm(w http.ResponseWriter, r *http.Request
 		return
 	}
 
+	mutex.Lock()
+	if _, exists := sessionMap[sessionID]; !exists {
+		sessionMap[sessionID] = &SessionState{IsArchiving: false}
+	}
+	mutex.Unlock()
+
 	// create unique user code assosiated with files
 	userCode := createUserCode()
 
@@ -91,10 +97,21 @@ func (app *application) getSnippet(w http.ResponseWriter, r *http.Request) {
 // homeGetFiles upload files to zip
 func (app *application) homeGetFiles(w http.ResponseWriter, r *http.Request) {
 
-	sessionID, userCode, err := app.getSessionValue(r)
+	sessionID, sessionCode, err := app.getSessionValue(r)
 	if err != nil {
 		app.serverError(w, err)
+		return
 	}
+
+	// check if the file is in Archiving status
+	mutex.Lock()
+	state := sessionMap[sessionID]
+	if state.IsArchiving {
+		mutex.Unlock()
+		app.serverErrorCode(w, fmt.Errorf("upload is prohibited while archive is in progress"), http.StatusForbidden)
+		return
+	}
+	mutex.Unlock()
 
 	var folderPathFull = filepath.Join(folderPath, sessionID)
 	app.logger.infoLog.Printf("create new folder %s", folderPathFull)
@@ -109,7 +126,7 @@ func (app *application) homeGetFiles(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	w.Write([]byte(userCode))
+	w.Write([]byte(sessionCode))
 }
 
 func (app *application) redirectToArchive(w http.ResponseWriter, r *http.Request) {
@@ -138,18 +155,32 @@ func (app *application) archive(w http.ResponseWriter, r *http.Request) {
 	sessionID, fileCode, err := app.getSessionValue(r)
 	if err != nil {
 		app.serverError(w, err)
+		return
 	}
+
+	// check if the file is in Archiving status
+	mutex.Lock()
+	state := sessionMap[sessionID]
+	if state.IsArchiving {
+		mutex.Unlock()
+		app.serverErrorCode(w, fmt.Errorf("archive operation is already in progress"), http.StatusConflict)
+		return
+	}
+
+	// Set the archiving flag
+	state.IsArchiving = true
+	sessionMap[sessionID] = state
+	mutex.Unlock()
 
 	// create zip archive with files
 	fileNameList, err := createArhive(sessionID, fileCode)
 	if err != nil {
 		app.serverError(w, err)
+		return
 	}
 
 	// select file lifetime by selected radio value
 	oneTimeDownload, lifeTime := selectLifeTime(r.FormValue("storageDuration"))
-
-	fmt.Println("archive lifeTime: ", lifeTime)
 
 	// collect file information
 	fullURL := getFullURL(r, fileCode)
@@ -157,6 +188,7 @@ func (app *application) archive(w http.ResponseWriter, r *http.Request) {
 	base64ImageData, err := createBase64ImageData(fullURL)
 	if err != nil {
 		app.serverError(w, err)
+		return
 	}
 
 	file := &models.File{
@@ -173,6 +205,7 @@ func (app *application) archive(w http.ResponseWriter, r *http.Request) {
 	fileJson, err := json.Marshal(file)
 	if err != nil {
 		app.serverError(w, err)
+		return
 	}
 
 	app.redisClient.HSet((app.getRedisPath(availablePath, fileCode)), fileInfoTitle, string(fileJson))
@@ -187,11 +220,20 @@ func (app *application) deleteOneFile(w http.ResponseWriter, r *http.Request) {
 	sessionID, _, err := app.getSessionValue(r)
 	if err != nil {
 		app.serverError(w, err)
+		return
 	}
 	fullPath := filepath.Join(folderPath, sessionID, fileName)
+
+	// Check if the file exists
+	if _, err := os.Stat(fileName); errors.Is(err, os.ErrNotExist) {
+		app.serverErrorCode(w, err, http.StatusNotFound)
+		return
+	}
+
 	err = os.Remove(fullPath)
 	if err != nil {
 		app.serverError(w, err)
+		return
 	}
 }
 
@@ -199,6 +241,7 @@ func (app *application) getUserCode(w http.ResponseWriter, r *http.Request) {
 	_, fileCode, err := app.getSessionValue(r)
 	if err != nil {
 		app.serverError(w, err)
+		return
 	}
 
 	w.Write([]byte(fileCode))
